requirements.txt
aerich==0.9.1
aiosmtplib==3.0.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgi-lifespan==2.1.0
asyncclick==8.2.2.2
asyncpg==0.30.0
bcrypt>=4.1.2,<5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
dictdiffer==0.9.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastapi-mail==1.5.0
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
git-filter-repo==2.47.0
google-auth==2.40.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
makefun==1.16.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.102.0
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pwdlib==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypika-tortoise==0.6.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.35.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tortoise-orm==0.25.1
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
